chore: create `toText` utility for RichText
