+++ packages/runtime/src/slate/index.ts
@@ -7,3 +7,4 @@ export * from './types'
 export * from './InlinePlugin'
 export * from './LinkPlugin'
 export * from './LocalChangesPlugin'
+export * from './utils'
