+++ packages/runtime/src/slate/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './toText'
